// src/types/supabase-rest.d.ts
export type components = {
    schemas: {
        devices: {
            /** Format: integer */
            battery_level?: number;
            /**
             * Format: uuid
             * @description Note:
             *     This is a Foreign Key to `rooms.id`.<fk table='rooms' column='id'/>
             */
            hoster_id?: string;
            /**
             * Format: uuid
             * @description Note:
             *     This is a Primary Key.<pk/>
             * @default gen_random_uuid()
             */
            id: string;
            /**
             * Format: timestamp with time zone
             * @default timezone('utc'::text, now())
             */
            inserted_at: string;
            /** Format: character varying */
            name: string;
            /**
             * Format: public.device
             * @enum {string}
             */
            type?: "display" | "keyboard" | "mouse" | "headset" | "webcam";
            /**
             * Format: timestamp with time zone
             * @default timezone('utc'::text, now())
             */
            updated_at: string;
        };
        rooms: {
            /** Format: character varying */
            address?: string;
            /**
             * Format: uuid
             * @description Note:
             *     This is a Foreign Key to `users.id`.<fk table='users' column='id'/>
             */
            booker_id?: string;
            /**
             * Format: uuid
             * @description Note:
             *     This is a Primary Key.<pk/>
             * @default gen_random_uuid()
             */
            id: string;
            /**
             * Format: timestamp with time zone
             * @default timezone('utc'::text, now())
             */
            inserted_at: string;
            /** Format: character varying */
            name: string;
            /**
             * Format: timestamp with time zone
             * @default timezone('utc'::text, now())
             */
            updated_at: string;
        };
        users: {
            /** Format: timestamp with time zone */
            created_at?: string;
            /** Format: character varying */
            email?: string;
            /** Format: text */
            first_name?: string;
            /**
             * Format: uuid
             * @description Note:
             *     This is a Primary Key.<pk/>
             */
            id?: string;
            /** Format: text */
            last_name?: string;
            /** Format: timestamp with time zone */
            updated_at?: string;
        };
    };
// ...
};